20-29 missing 25

+++ 03_020_imprecision.c
@@ -2,12 +2,12 @@
 #include <stdlib.h>
 
 int main() {
-    float resultat = 1.;
+    /*float resultat = 1.;*/
+    double resultat = 1.;
     float ajout = 1e-9;
     printf("resultat = %.15f\n", resultat);
     printf("on ajoute  %.15f\n", ajout);
     resultat += ajout;
     printf("resultat = %.15f\n", resultat);
-    /* Pourquoi resultat ne change pas ? */
     exit(EXIT_SUCCESS);
 }
